Implement TimePickerDialog for reinder preference and refactor settings package to preference
