+++ app/src/main/res/layout/activity_preference.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".settings.PreferenceActivity">
+    tools:context=".preference.PreferenceActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
