+++ app/src/main/res/layout/activity_settings.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.isscroberto.dailyprayerandroid.settings.SettingsActivity">
+    tools:context="com.isscroberto.dailyprayerandroid.preference.SettingsActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
